Move components to src in @commons-ui/next
