+++ packages/commons-ui-next/package.json
@@ -2,7 +2,7 @@
   "name": "@commons-ui/next",
   "version": "2.0.0-alpha.0",
   "private": false,
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "jest",
     "lint": "eslint --fix './'",
