+++ packages/commons-ui-next/src/Link/index.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Link from "./Link";
+import Link from ".";
 
 export default {
   title: "Components/Link",
