+++ packages/commons-ui-next/src/Link/index.test.js
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import React from "react";
 
-import Link from "./Link";
+import Link from ".";
 
 describe("Link", () => {
   it("renders unchanged", () => {
