+++ packages/commons-ui-next/src/index.js
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+
+export { default as Link } from "./Link";
